# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PdhFeedback SaaS Developer REST API",
    "version": "1.0.0",
    "description": "API สำหรับเชื่อมต่อระบบประเมินความพึงพอใจ PdhFeedback กับระบบสารสนเทศโรงพยาบาล (HIS), CRM, POS และเว็บไซต์ภายนอกแบบ Multi-tenant SaaS พร้อมระบบสิทธิ์ Scoped API Keys และ Idempotency",
    "contact": {
      "name": "PdhFeedback Engineering Team",
      "email": "support@pdhfeedback.local"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Environment Server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/api/v1/service-points": {
      "get": {
        "summary": "เรียกดูรายชื่อจุดบริการทั้งหมด",
        "description": "ดึงข้อมูลจุดบริการ แผนก หรือสาขาที่เปิดใช้งานขององค์กร",
        "operationId": "getServicePoints",
        "security": [{ "BearerAuth": ["service_points:read"] }],
        "responses": {
          "200": {
            "description": "รายการจุดบริการ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "servicePoints": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/ServicePoint" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/UnauthorizedError" },
          "403": { "$ref": "#/components/responses/ForbiddenError" }
        }
      }
    },
    "/api/v1/surveys": {
      "get": {
        "summary": "เรียกดูแบบประเมินและเวอร์ชันที่เปิดใช้งาน",
        "description": "ดึงโครงสร้างแบบประเมิน เวอร์ชันที่ Publish แล้ว และรหัส Publication สำหรับนำไปฝัง",
        "operationId": "getSurveys",
        "security": [{ "BearerAuth": ["surveys:read"] }],
        "responses": {
          "200": {
            "description": "รายการแบบประเมินและโครงสร้างคำถาม",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "surveys": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Survey" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invitations": {
      "post": {
        "summary": "สร้างลิงก์คำเชิญประเมินเฉพาะรายการบริการ (One-time Invitation)",
        "description": "สร้างลิงก์สำหรับส่งต่อให้ผู้รับบริการทาง SMS/Line/Queue Ticket สามารถระบุ Idempotency-Key ใน Header เพื่อป้องกันการสร้างซ้ำ",
        "operationId": "createInvitation",
        "security": [{ "BearerAuth": ["invitations:write"] }],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": { "type": "string" },
            "description": "รหัสป้องกันการทำรายการซ้ำ เช่น UUID หรือ Visit ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["publicationId"],
                "properties": {
                  "publicationId": { "type": "string", "description": "รหัส Publication หรือ publicCode" },
                  "servicePointId": { "type": "string", "description": "รหัสจุดบริการ (ไม่บังคับ)" },
                  "expiresInDays": { "type": "integer", "default": 7, "minimum": 1, "maximum": 30 },
                  "externalReference": { "type": "string", "description": "รหัสอ้างอิง Opaque (ห้ามเก็บ HN หรือชื่อสกุล)" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "คำเชิญถูกสร้างสำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "invitationId": { "type": "string" },
                    "invitationUrl": { "type": "string", "format": "uri" },
                    "expiresAt": { "type": "string", "format": "date-time" },
                    "externalReference": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invitations/{invitationId}/revoke": {
      "post": {
        "summary": "เพิกถอนคำเชิญ",
        "description": "ระงับการใช้งานลิงก์คำเชิญก่อนหมดอายุ",
        "operationId": "revokeInvitation",
        "security": [{ "BearerAuth": ["invitations:write"] }],
        "parameters": [
          {
            "name": "invitationId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "เพิกถอนสำเร็จ" }
        }
      }
    },
    "/api/v1/reports/summary": {
      "get": {
        "summary": "สรุปสถิติคะแนนความพึงพอใจ (Aggregate Metrics)",
        "description": "คำนวณ CSAT, NPS, คะแนนเฉลี่ย และจำนวนคำตอบรวมตามช่วงเวลา (ไม่เปิดเผยข้อมูลผู้ตอบ)",
        "operationId": "getReportSummary",
        "security": [{ "BearerAuth": ["reports:read"] }],
        "parameters": [
          { "name": "range", "in": "query", "schema": { "type": "string", "enum": ["7d", "30d", "90d", "1y", "custom"], "default": "30d" } },
          { "name": "startDate", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "endDate", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "servicePointId", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "สรุปตัวเลขสถิติ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "metrics": {
                      "type": "object",
                      "properties": {
                        "totalResponses": { "type": "integer" },
                        "csat": { "type": "number", "description": "เปอร์เซ็นต์คะแนน 4-5 ดาว" },
                        "averageScore": { "type": "number", "description": "คะแนนเฉลี่ย 1.0 - 5.0" },
                        "nps": { "type": "number", "description": "Net Promoter Score (-100 to 100)" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/ratings": {
      "get": {
        "summary": "การแจกแจงคะแนนดาว 1–5",
        "description": "แสดงจำนวนและเปอร์เซ็นต์ของคะแนนแต่ละระดับดาว 1-5",
        "operationId": "getReportRatings",
        "security": [{ "BearerAuth": ["reports:read"] }],
        "responses": {
          "200": { "description": "การแจกแจงระดับคะแนนดาว" }
        }
      }
    },
    "/api/v1/usage": {
      "get": {
        "summary": "ตรวจสอบโควตาและการใช้งาน",
        "description": "อ่านจำนวนคำตอบที่ใช้ไปในรอบเดือน โควตาสูงสุด และขีดจำกัดตามแพ็กเกจ",
        "operationId": "getUsage",
        "security": [{ "BearerAuth": ["usage:read"] }],
        "responses": {
          "200": { "description": "ข้อมูลการใช้งานและโควตา" }
        }
      }
    },
    "/api/v1/public/surveys/{publicationId}": {
      "get": {
        "summary": "อ่านโครงสร้างแบบประเมินสำหรับ Embed (Public)",
        "description": "เข้าถึงได้โดยไม่ต้องใช้ API Key สำหรับใช้เรนเดอร์ใน widget/iframe",
        "operationId": "getPublicSurvey",
        "security": [],
        "parameters": [
          { "name": "publicationId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "โครงสร้างแบบประเมินสาธารณะ" },
          "404": { "description": "ไม่พบแบบประเมินหรือปิดรับคำตอบ" }
        }
      }
    },
    "/api/v1/public/surveys/{publicationId}/responses": {
      "post": {
        "summary": "ส่งผลการประเมิน (Public Submission)",
        "description": "บันทึกผลการประเมินรูปดาว 1–5 และคำตอบเพิ่มเติม พร้อมระบบป้องกันการส่งซ้ำ (Idempotency)",
        "operationId": "submitPublicResponse",
        "security": [],
        "parameters": [
          { "name": "publicationId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "Idempotency-Key", "in": "header", "required": false, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["overallRating"],
                "properties": {
                  "overallRating": { "type": "integer", "minimum": 1, "maximum": 5, "description": "คะแนนดาว 1-5" },
                  "npsScore": { "type": "integer", "minimum": 0, "maximum": 10 },
                  "commentText": { "type": "string", "maxLength": 1000 },
                  "invitationToken": { "type": "string" },
                  "answers": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["questionId"],
                      "properties": {
                        "questionId": { "type": "string" },
                        "ratingValue": { "type": "integer" },
                        "textValue": { "type": "string" },
                        "selectedOptions": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "บันทึกคำตอบสำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "responseId": { "type": "string" },
                    "submittedAt": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "429": { "description": "เกินอัตราการเรียกหรือโควตาประจำเดือนเต็ม" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key (pdh_live_...)",
        "description": "ระบุ API Key ที่ได้รับจาก Integration Center"
      }
    },
    "schemas": {
      "ServicePoint": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "code": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "publicCode": { "type": "string" }
        }
      },
      "Survey": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "slug": { "type": "string" },
          "currentVersion": { "type": "integer" }
        }
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "ไม่ได้รับอนุญาต (API Key ไม่ถูกต้องหรือไม่ได้ระบุ)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": false },
                "error": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "string", "example": "UNAUTHORIZED" },
                    "message": { "type": "string" },
                    "requestId": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "ไม่มีสิทธิ์เข้าถึง (Scope ไม่เพียงพอ หรือแพ็กเกจไม่ครอบคลุม)"
      }
    }
  }
}